Deliverable 6
So for this deliverable I set my local server to ask permission to have that persons data for their primary calendar. I haded the google chrome api to do that. Then it asked for an access token and in the outh.php document I did that then I created the html page in which these thee will take you to.